feat: refactor date visibility
